+++ src/card/cards/stale_followups.rs
@@ -88,3 +88,4 @@ pub const WALL_OF_WONDER: CardDefinitionId = CardDefinitionId(1732);
 pub const INFERNAL_MEDUSA: CardDefinitionId = CardDefinitionId(1733);
 pub const VENOM: CardDefinitionId = CardDefinitionId(1734);
 pub const DAMPING_FIELD: CardDefinitionId = CardDefinitionId(1735);
+pub const ABU_JAFAR: CardDefinitionId = CardDefinitionId(1736);